requirements.txt
pytest==8.4.1
requests==2.32.4
jsonschema==4.24.0
python-dotenv
